+++ src/resources/tasks/task.router.js
@@ -1,31 +1,51 @@
 const Router = require('@koa/router');
 const tasksService = require('./task.service');
+const {uuidValidator, entityExistValidator} = require('../../common/utils/common.validator');
 
 const router = new Router({
   prefix: '/boards/:boardId/tasks'
 });
 
 router.get('/', ctx => {
+  const {boardId} = ctx.params;
+
+  if (uuidValidator(boardId, ctx)) {
+    return;
+  }
+
+  if (entityExistValidator(boardId, 'board', ctx)) {
+    return;
+  }
+
   const tasks = tasksService.getAll();
 
   ctx.body = tasks.map(task => (task.toResponse()));
 });
 
 router.get('/:id', ctx => {
-  const {id} = ctx.params;
+  const {boardId, id} = ctx.params;
 
-  if (tasksService.getTaskById(id)) {
-    ctx.body = tasksService.getTaskById(id).toResponse();
-
+  if (uuidValidator(boardId, ctx) || uuidValidator(id, ctx)) {
     return;
   }
 
-  ctx.response.status = 404;
-  ctx.body = "Board not found.";
+  if (entityExistValidator(boardId, 'board', ctx) || entityExistValidator(id, 'task', ctx)) {
+    return;
+  }
+
+  ctx.body = tasksService.getTaskById(id).toResponse();
 });
 
 router.post('/', ctx => {
   const {boardId} = ctx.params;
+
+  if (uuidValidator(boardId, ctx)) {
+    return;
+  }
+
+  if (entityExistValidator(boardId, 'board', ctx)) {
+    return;
+  }
 
   let task = ctx.request.body;
   task.boardId = boardId;
@@ -38,6 +58,14 @@ router.post('/', ctx => {
 router.put('/:id', ctx => {
   const {boardId, id} = ctx.params;
 
+  if (uuidValidator(boardId, ctx) || uuidValidator(id, ctx)) {
+    return;
+  }
+
+  if (entityExistValidator(boardId, 'board', ctx) || entityExistValidator(id, 'task', ctx)) {
+    return;
+  }
+
   ctx.request.body.boardId = boardId;
   const task = tasksService.updateTask(id, ctx.request.body);
 
@@ -46,7 +74,15 @@ router.put('/:id', ctx => {
 });
 
 router.delete('/:id', ctx => {
-  const {id} = ctx.params;
+  const {boardId, id} = ctx.params;
+
+  if (uuidValidator(boardId, ctx) || uuidValidator(id, ctx)) {
+    return;
+  }
+
+  if (entityExistValidator(boardId, 'board', ctx) || entityExistValidator(id, 'task', ctx)) {
+    return;
+  }
 
   tasksService.deleteTask(id);
